clients/ui-streamlit/pyproject.toml
[project]
name = "client-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.18",
    "langchain>=0.3.20",
    "langchain-anthropic>=0.1.1",
    "langchain-google-genai>=2.1.2",
    "langchain-mcp-adapters>=0.0.7",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.0.3",
    "langgraph>=0.3.30",
    "nest-asyncio>=1.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44",
    "streamlit-aggrid>=0.3.4.post3",
    "streamlit-option-menu>=0.3.6",
]
